+++ src/index.ts
@@ -1,12 +1,12 @@
 import {app} from "./app";
-import { cadastrarDocente } from "./endpoints/cadastrarDocente";
-import { cadastrarEstudante } from "./endpoints/cadastrarEstudante";
-import { buscarEstudante } from "./endpoints/buscarEstudante";
+import { cadastrarDocente } from "./endpoints/docentes/cadastrarDocente";
+import { cadastrarEstudante } from "./endpoints/estudantes/cadastrarEstudante";
+import { buscarEstudante } from "./endpoints/estudantes/buscarEstudante";
 import { criarTurma } from "./endpoints/criarTurma";
 import { testeInicial } from "./endpoints/testeInicial";
-import { mudarEstudante } from "./endpoints/mudarEstudante";
-import { buscarDocente } from "./endpoints/buscarDocentes";
-import { mudarDocente } from "./endpoints/mudarDocente";
+import { mudarEstudante } from "./endpoints/estudantes/mudarEstudante";
+import { buscarDocente } from "./endpoints/docentes/buscarDocentes";
+import { mudarDocente } from "./endpoints/docentes/mudarDocente";
 
 app.get("/", testeInicial)
